requirements.txt
 
fastapi
uvicorn
pydantic
google-generativeai
newspaper3k
trafilatura
beautifulsoup4
lxml
spacy
requests
aiohttp
python-multipart
jinja2
markdown
python-dotenv
typing-extensions
httpx
selenium
webdriver-manager
readability-lxml
validators
urllib3
html2text
dateparser
langdetect
cssselect